bfs: Using a "long" to hold the chunk size is not sensible.
Former-commit-id: 32c57a1015fef1672605d9c9ff4032ae1631f25d

+++ bfs/trunk/src/bfs/BFSFileInfo.java
@@ -35,7 +35,7 @@ public class BFSFileInfo implements Serializable, Comparable<BFSFileInfo> {
 		return Conf.getRepFactor();
 	}
 
-	public long getChunkSize() {
+	public int getChunkSize() {
 		return Conf.getChunkSize();
 	}
